feat(components-js/react): move testing-library utils to components-js and reexport for react | bh | #2203
